pyproject.toml
[tool.poetry]
name = "ergochat_auth"
version = "0.1.0"
description = "Script to authenticate ergo users against a MongoDB database."
authors = ["DeadOcean <admin@suprachat.net>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.19.2"
pymongo = "^4.0.1"
Werkzeug = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
